Record a FAILING test which occurs when rules don't take actions
The 'allowed?' rule used to check access seems to return a permission
that doesn't relate to the action, in situations where numerous
actions are passed into the access control check. Therefore, more
possible actions are returned than we expect (4 versus 3).

In the test, we don't expect the subject to have access to
put-session-scope and yet that is what the algorithm returns.