Rename concept of actions to operations
Site operations map almost directly onto an operation in
OpenAPI. A GraphQL operation, for the most part, maps to a single Site
operation or set of Site operations.

There doesn't seem to be sufficient justification to use a different
word for a similar concept, preventing users from leverage
understanding from OpenAPI and GraphQL. The minor technical
differences can be explained when necessary at a more advanced level.

The term 'operations' is used for the same or similar concept in
https://en.wikipedia.org/wiki/Role-based_access_control