Disable Malli instrumentation
Hunting down the reason why authentication isn't promoting a
session. This bug appears to have gone away.

Without having got to the bottom of the issue, my experiments were
suggesting that having 'put' a subject to the database, and after
having awaited for that transaction, the subject is not yet available
in the next request, causing a bounce back to the login screen.

I even tried downgrading XTDB (didn't change anything).

This might have something to do with misconfiguration, for example,
the session scope used by the authenticator and application
authorizer.