Move users/user-identities from auth.example.org to data.example.org
We want to access users and user-identities directly via an API. The
URIs of users/user-identities is part of the API. We want the URIs to
be something we can send requests (with bearer tokens) rather than be
accessible via session cookies. Sessions should only be used for
authentication in the process of acquiring an access tokens. In other
words, where data is concerned, always prefer access tokens over
session tokens.

Access tokens are better than session tokens in that they have more
features, such as limiting scope.