Use JWTs for access tokens
This allows us to generate the correct CORS origin plus allows
front-end JavaScript to know which access-token should be sent to
which origin, by looking at the 'aud' claim.

Note, these JWTs are not yet signed but are still secure because they
have a hard-to-guess jti claim and the entire JWT is looked up in the
database.