Separate out installation of authorization and token endpoints
We want to install the token endpoint in the bootstrap, so that we can
use the client_credentials flow to allow insite to carry on the
configuration process remotely.

However, we don't want to install the authorization endpoint in the
bootstrap, because it requires certain user decisions to be
made (e.g. which authentication provider to use).