Use a keypair 'alias' to remove coupling
`/oauth/authorize` and `/oauth/token` resources depend on a keypair,
as both need to sign JWTs they might generate. Where
`/oauth/authorize` generates JWT for the implicit flow, `/oauth/token`
generates tokens for `authorization_code`, `password` and
`client_credentials` flows.

If we tightly bind one or both of these resources to a keypair, it
makes it more difficult to rotate the keypair in future.

This commit introduces a strategy whereby when a new keypair is
generated (using the `/operations/generate-keypair` operation), a
keypair is generated and written as two resources. The first is a new
resource, with the kid in the URI. The second is to write (and
potentially overwrite) the `/keypairs/current-signing-key` resource.

When a new keypair is generated, the old signing key is effectively
rotated to the new keypair, which leaving older keypairs available for
verifying signatures of JWTs that have been signed prior to the
rotation.

Currently the `/keypairs/current-signing-key` does not expire, but it
would be possible to more tightly control its validity to, say, 3
months. This would mean that after the expiry period, the system
wouldn't be able to sign new JWTs. Possibly a less dramatic keypair
expiry might be desirable.