Move from bash to babashka for sitectl/site
I noticed that only first names were appearing in a fullname
attribute, when using the 'site' tool. After quite a lot of debugging,
this issue was isolated to the way that the bash scripts were
processing arguments (quoted arguments were becoming unquoted).

I was not successful in adopting any of the bash solutions here:
https://stackoverflow.com/questions/10835933/how-can-i-preserve-quotes-in-printing-a-bash-scripts-arguments

Therefore, I've moved to babashka-calling-babashka to eliminate the
insanity that isarchaic bash syntax.