Move assign-role operation back to auth.example.org
The reason it was moved in the first place was to protect it with a
bearer token. However, resources can be individually protected, rather
than having to be grouped under a particular host.

The operations for managing Site's users, roles and internal
permissions should remain with the Authorization Server for now.