Serve code-on-demand to a 401 response
This is the initial work to serve authorized HTML views to browsers as
part of a range of improvements to the developer experience.

It should be possible to work with Site reasonably well without
requiring external applications. The browser is a powerful tool to
view HTML representations. However, for pages requiring authorization,
we must give it some help, in the form of some attached JavaScript
that will initiate a flow to acquire an access token which can then be
used to retrieve protected information and update the relevant content
areas of the page.