Prepare all resources in prepare phase
When the install-resources operation is called, each installer should
be prepared to discover any errors prior to the transaction function
being called.

Prior to this commit, it was found that a validation failure in an
installer could result in a SCI error message containing source
information in a volatile. This indirectly caused a perpetual Nippy
serialization error in XTDB, with the only fix being to delete and
re-init the database!

This issue itself hasn't been fixed by this commit. Rather, in the
install-resources operation, the installer-seq->tx-ops is called in
the prepare phase rather than the transact phase. This means that SCI
errors which might occur in the prepare phase on any of the resources
within the bundle will cause the operation to abort before its
transaction function is run.

If there are any SCI errors in a transact phase, then these will still
cause the serious database corruption issue. This will further
investigation and a future fix. However, we may be upgraing to XTDBv2
in due course, where the fix may not be necessary, or may be
different.