Refactor to replace 'converge!' with bundles
There were a few remaining places where the old pre-bundle `converge!`
function was being called by tests.

This commit also fixes a minor issue whereby ambiguity between
user-role assignments and application-role assignments has been
removed. This ambiguity meant that it wouldn't be possible to have a
user with the same name as an application client-id both have a
particular role, due to a clash in the URI namespace.