Replace juxt.site/base-resource with juxt.site/variant
This introduces a new entry into the request
context (:juxt.site/variant) which, if set, contains the
variant (resource) which is a variant of the entry in
:juxt.site/resource.

In other words, we no longer lose track of the 'base-resource', which
can be useful in future for defaulting behaviour.

Most importantly, it introduces the same terminology as used in HTTP
specifications and hopefully makes Site's content negotiation easier
to reason about.