Remove access to *ctx* in transaction functions
This is a preparatory step towards being more minimal in what we put
on the transaction log when we call each transaction function.

Currently, the entire request context is placed on the transaction
log, which contributes significantly to the overall size of the
transaction on the transaction log.

This commit only modifies a few operations to avoid using *ctx* in
their transact phases.

A future commit may modify the operations code to be more much
selective about what goes into each transaction.