Allow applications to call certain operations
This relaxes the constraint that site-cli can only do a limited amount
of setup. Even though it is desirable (for audit logging) to create a
user as early as possible, for 'site bootstrap' to do more work we
need to allow the site-cli application to do more.

We can revert this simply by removing the assignment of the SiteAdmin
role to the site-cli application.