Support writing OpenAPI docs in openapi.edn
It is true that we want to move towards making Site a JSON-by-default
product, but that should not exclude supporting native Clojure formats
where it can be done invisibly to other users. In this case, it is
possible to support an additional content type (application/edn)
without deviating from the JSON-by-default policy.