Complete removal of https://meta.juxt.site/types/resource
The purpose of removing this is to disambiguate resources with no
type, and resources with the type of
https://meta.juxt.site/types/resource. This ensures that resources
with no type, are treated as plain resources, rather than their
behaviour being undefined.

The final usage of https://meta.juxt.site/types/resource has been in
the area of selmer templating. It is important that templates cannot
include resources with certain types, as this might allow subjects
with the ability to put templates to write templates in such a way as
to include sensitive information. Now, a template might be included by
another template, with render-file, must be marked with the specific
type of https://meta.juxt.site/types/resource. This has meant that a
new bundle (juxt/site/selmer-templating) has been added, along with an
operation and associated permissions.