Allow the dynamic creation of resources
Until now, resources are installed by creating bundles (stored in a
git repo) and installing them. While this is a tactical necessity for
bootstrapping the system, and for supporting git-based use-cases, this
does not realise the full potential of Site as a fully-dynamic
programmable system.

Ideally, resources are created in a running Site instance dynamically,
with 'source code control' being provided by the bitemporality of the
database. OpenAPI documents and bundles should be composed from such
resources, incrementally. Bundles can then be published via a Site
instance, and exported as files if necessary (where a bundle must be
exported from Site instance A and installed into Site instance B, and
where B does not have network connectivity to A).